backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.3
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2023.11.17
pillow==10.1.0
aiofiles==23.2.1
requests==2.31.0